# api/docs/swagger.yaml
basePath: /
definitions:
  dto.BatchRequest:
    properties:
      correlation_id:
        type: string
      original_url:
        type: string
    type: object
  dto.BatchResponse:
    properties:
      correlation_id:
        type: string
      short_url:
        type: string
    type: object
  dto.ShortenRequest:
    properties:
      url:
        type: string
    type: object
  dto.ShortenResponse:
    properties:
      result:
        type: string
    type: object
  dto.UserURL:
    properties:
      original_url:
        type: string
      short_url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис сокращения ссылок.
  title: Shortener API
  version: "1.0"
paths:
  /{id}:
    get:
      description: Получает оригинальный URL по его короткому идентификатору и делает
        перенаправление.
      parameters:
      - description: Короткий идентификатор ссылки
        in: path
        name: id
        required: true
        type: string
      responses:
        "307":
          description: Temporary Redirect
          schema:
            type: string
        "404":
          description: URL not found
          schema:
            type: string
        "410":
          description: URL deleted
          schema:
            type: string
      summary: Перенаправление по короткой ссылке
      tags:
      - redirect
  /api/shorten:
    post:
      consumes:
      - application/json
      description: Принимает оригинальный URL в формате JSON и возвращает короткую
        ссылку.
      parameters:
      - description: Данные для сокращения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ShortenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Короткая ссылка создана
          schema:
            $ref: '#/definitions/dto.ShortenResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "409":
          description: Ссылка уже существует
          schema:
            $ref: '#/definitions/dto.ShortenResponse'
        "500":
          description: internal error
          schema:
            type: string
      summary: Создать короткую ссылку (JSON API v1.3)
      tags:
      - urls
  /api/shorten/batch:
    post:
      consumes:
      - application/json
      description: Принимает массив исходных URL и возвращает массив сокращённых ссылок
      parameters:
      - description: Список ссылок для сокращения
        in: body
        name: input
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.BatchRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.BatchResponse'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      summary: Сократить ссылки пачкой
      tags:
      - urls
  /api/user/urls:
    delete:
      consumes:
      - application/json
      description: Принимает список коротких ссылок пользователя и отправляет их на
        асинхронное удаление
      parameters:
      - description: Список коротких ссылок для удаления
        in: body
        name: input
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "202":
          description: Запрос принят на обработку
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            type: string
      summary: Удалить сокращённые ссылки пачкой
      tags:
      - urls
    get:
      description: Возвращает список всех сокращённых ссылок, принадлежащих текущему
        пользователю
      produces:
      - application/json
      responses:
        "200":
          description: Список ссылок
          schema:
            items:
              $ref: '#/definitions/dto.UserURL'
            type: array
        "204":
          description: Нет ссылок
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Получить все сокращённые ссылки пользователя
      tags:
      - urls
  /ping:
    get:
      description: Отправляет ping к базе данных. Если соединение установлено, возвращает
        200 OK.
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: db unavailable" или "db ping error
          schema:
            type: string
      summary: Проверка подключения к базе данных
      tags:
      - health
swagger: "2.0"